map features + fixes
- add flag to override env_fog distance adjustment for special cases (mystic_radar_v1)
- robogrunt can use rpg/sniper (shockraid_jungle)
- fix crash precaching an empty string
- allow train to move from/to the same path_corner (hl_u12_a5)
- fix warnings
- hwgrunt stuck in a cover/spinup loop (svencontra2)
- fix ambient_generic direction wrong when far from the world origin

+++ dlls/monster/CBaseMonster.cpp
@@ -3032,10 +3032,10 @@ void CBaseMonster::HandleAnimEvent(MonsterEvent_t* pEvent)
 		if (flash.sprite) {
 			Vector origin, angles;
 
-			if (flash.attachment != -1) {
+			if (flash.attachment != (uint8_t)-1) {
 				GetAttachment(flash.attachment, origin, angles);
 			}
-			else if (flash.bone != -1) {
+			else if (flash.bone != (uint8_t)-1) {
 				GetBonePosition(flash.bone, origin, angles);
 				origin = origin + flash.offset;
 			}
